refactor: 파일대소문자 수정,기타 수정(#71)
