+++ src/index.ts
@@ -10,7 +10,7 @@ import { dirname } from "path";
 import { Taskl, Task, TasklOptions } from "taskl";
 import { legacyFormatCommitMessage } from "./legacyFormatCommitMessage.js";
 import { detectCommitMode } from "./detectCommitMode.js";
-import { isValidOpenAIKey } from "./isValidOpenAIkey.js";
+import { isValidOpenAIKey } from "./isValidOpenAIKey.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -150,7 +150,9 @@ async function generateCommitMessage(commitMsgFile: string): Promise<void> {
   }
 
   const result = await isValidOpenAIKey(process.env.OPENAI_API_KEY);
-  result! && console.warn("ERROR");
+  if (result) {
+    console.warn("ERROR: is Not Valid API KEY");
+  }
 
   const tasks: Task[] = [
     {
